requirements.txt
absl-py==2.1.0
accelerate==1.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
anyio==4.8.0
attrs==25.1.0
autocommand==2.2.2
backports.tarfile==1.2.0
bing-image-downloader==1.1.2
bitsandbytes==0.42.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
datasets==3.3.2
dill==0.3.8
docopt==0.6.2
duckduckgo_search==7.4.5
einops==0.8.1
filelock==3.17.0
flash_attn==2.7.4.post1
frozenlist==1.5.0
fsspec==2024.12.0
grpcio==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.0.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.5
lxml==5.3.1
Markdown==3.7
MarkupSafe==3.0.2
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
num2words==0.5.14
numpy==2.2.3
packaging==24.2
pandas==2.2.3
peft==0.14.0
pillow==11.1.0
platformdirs==4.2.2
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scipy==1.15.2
setuptools==75.8.0
simple-image-download==0.2
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tokenizers==0.21.0
tomli==2.0.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers.git@92c5ca9dd70de3ade2af2eb835c96215cc50e815
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.43.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.19.2
